Refactored services.

+++ trms/src/main/java/com/revature/trms/services/EmployeeTypeServiceImpl.java
@@ -3,8 +3,10 @@ package com.revature.trms.services;
 import java.util.List;
 
 import com.revature.trms.daos.EmployeeTypeDAO;
+import com.revature.trms.exceptions.IllegalParameterException;
 import com.revature.trms.pojos.EmployeeType;
 import com.revature.trms.utilities.DAOUtilities;
+import com.revature.trms.utilities.LogUtilities;
 
 public class EmployeeTypeServiceImpl extends BaseService implements EmployeeTypeService {
 
@@ -15,34 +17,51 @@ public class EmployeeTypeServiceImpl extends BaseService implements EmployeeType
 	}
 
 	@Override
-	public boolean addEmployeeTypesToEmployee(Integer employeeId, List<EmployeeType> employeeTypeIds) {
+	public boolean addEmployeeTypesToEmployee(Integer employeeId, List<EmployeeType> employeeTypeIds)
+			throws IllegalParameterException {
+		LogUtilities.trace("addEmployeeTypesToEmployee");
+
+		if (employeeId == null) {
+			throw new IllegalParameterException("addEmployeeTypesToEmployee - employeeId should not be empty");
+		}
+
 		return employeeTypeDao.addEmployeeTypesToEmployee(employeeId, employeeTypeIds);
 	}
 
 	@Override
-	public boolean removeEmployeeTypesFromEmployee(Integer employeeId) {
+	public boolean removeEmployeeTypesFromEmployee(Integer employeeId) throws IllegalParameterException {
+		LogUtilities.trace("removeEmployeeTypesFromEmployee");
+
 		if (employeeId == null) {
-			throw new IllegalArgumentException("EmployeeId should not be empty.");
+			throw new IllegalParameterException("removeEmployeeTypesFromEmployee - employeeId should not be empty");
 		}
-		
+
 		return employeeTypeDao.removeEmployeeTypesFromEmployee(employeeId);
 	}
 
 	@Override
-	public List<EmployeeType> getEmployeeTypesForEmployee(Integer employeeId) {
+	public List<EmployeeType> getEmployeeTypesForEmployee(Integer employeeId) throws IllegalParameterException {
+		LogUtilities.trace("getEmployeeTypesForEmployee");
+
 		if (employeeId == null) {
-			throw new IllegalArgumentException("EmployeeId should not be empty.");
+			throw new IllegalParameterException("getEmployeeTypesForEmployee - employeeId should not be empty");
 		}
-		
+
 		return employeeTypeDao.getEmployeeTypesForEmployee(employeeId);
 	}
 
 	@Override
-	public boolean updateEmployeeTypesToEmployee(Integer employeeId, List<EmployeeType> employeeTypeIds) {
+	public boolean updateEmployeeTypesToEmployee(Integer employeeId, List<EmployeeType> employeeTypeIds) throws IllegalParameterException {
+		LogUtilities.trace("updateEmployeeTypesToEmployee");
+
+		if (employeeId == null) {
+			throw new IllegalParameterException("updateEmployeeTypesToEmployee - employeeId should not be empty");
+		}
+		
 		if (employeeTypeDao.removeEmployeeTypesFromEmployee(employeeId)) {
 			return employeeTypeDao.addEmployeeTypesToEmployee(employeeId, employeeTypeIds);
 		}
-		
+
 		return false;
 	}
